% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{bright}
\alias{bright}
\title{Checking For Bright Images}
\usage{
bright(
  imageMatrix,
  beta0 = -2.774327,
  beta1 = 51.91687,
  cutoffProbability = 0.5,
  NADefault = 0
)
}
\arguments{
\item{imageMatrix}{imageMatrix points representing the brightness of that pixel}

\item{beta0}{The intercept of a logistic regression default is for magnetograms}

\item{beta1}{The slope of the logistic regression default is for magnetograms}

\item{cutoffProbability}{The probability cut off for the decision of an imageMatrix being bright}

\item{NADefault}{The default value set to NA's in the matrix}
}
\value{
The decision of the image being bright (based off of the cutoffProbability)
}
\description{
Using a logistic regression to find coefficients, uses the result to compute a probability
of the imageMatrix being #'bright (392 observations) with McFadden Pseudo- $R^2$ Value 0.2573778
}
