% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{create_trace}
\alias{create_trace}
\title{Create Trace}
\usage{
create_trace(
  traceMatrix,
  start,
  end,
  topEnv,
  bottomEnv,
  thresh = 5,
  MARange = 6,
  region = 2,
  loopNumber = 4
)
}
\arguments{
\item{traceMatrix}{The matrix of an isolated trace}

\item{start}{Start line for the trace}

\item{end}{End line for the trace}

\item{topEnv}{The top envelope for the isolated trace Matrix}

\item{bottomEnv}{The bottom envelope for the isolated trace Matrix}

\item{thresh}{The threshold for the diff() as a cut off for allowed spikes}

\item{MARange}{The amount in each direction that the moving average will look at (added to the region value)}

\item{region}{The region that is smoothed by the moving average}

\item{loopNumber}{How many times it will go through this process, (to catch the new peaks from smoothing)}
}
\value{
Matrix of the line, corrected for the matrix
}
\description{
Takes a top and bottom trace on an isolated image and creates line, with MA smoothing
}
