% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.allowed_edge_distance}
\alias{.allowed_edge_distance}
\title{Maximum/Minimum distance(percentage) from Edge}
\usage{
.allowed_edge_distance(rowSums, percentage, percentageLeftSide = NULL)
}
\arguments{
\item{rowSums}{The sum of the rows of the matrix}

\item{percentage}{Percentage of the picture that you would like removed from edge
if percentageLeftSide is unspecified uses percentage for both i.e left = 1 - right}

\item{percentageLeftSide}{the left percentage that you want to remove(default = NULL)}
}
\value{
vector of (rightSideDistance, leftSideDistance)
}
\description{
Takes a percentage and calculates the cutoff value for a
function to look for something.  Used for find_peaks
}
