% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.bottom_env}
\alias{.bottom_env}
\title{Bottom Envelope}
\usage{
.bottom_env(rolledImage, max_roc = 25, sepDist = 10, maxNoise = 100)
}
\arguments{
\item{rolledImage}{Image that has been put through the .roll_image()}

\item{max_roc}{maximum rate of change allowed between two pixels on the line before deemed as noise}

\item{sepDist}{how far you want the envelope to be below the line you are tracing}

\item{maxNoise}{the length of creating points allowed before considered to be off of the trace}
}
\value{
vector of points for a line, in the correct scaling for the rolled image.
can do nrow(imageMatrix) - bottomcut + Envelope change the scaling
}
\description{
Bottom bound for a set of horizontal lines, follows contour of the top line and corrects for un-wanted noise
}
