% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.bottom_image_cut}
\alias{.bottom_image_cut}
\title{Bottom Image Cut}
\usage{
.bottom_image_cut(
  imageMatrix,
  percentFromEdge,
  percentEdgeForLeft = NULL,
  shortestAllowedSeqOfZeros = 50
)
}
\arguments{
\item{imageMatrix}{Imported image into matrix form, can use tiff_import()}

\item{percentFromEdge}{used in find_peaks if you know there wont be a peak
in a region}

\item{percentEdgeForLeft}{passed into find peaks, if not specified, uses
percentFromEdge for both left and right sides, if specified, percentFromEdge
is defaulted to just the right side of the plot}

\item{shortestAllowedSeqOfZeros}{smallest gap allowed to be found to consider
the trace not intersecting the timing marks}
}
\value{
value of bottom cut that should be removed
}
\description{
attempts to find a gap between the traces and timing marks, if found it will
 return the point in which the image can be trimmed removing the timing marks
}
