% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.bottom_upper_env}
\alias{.bottom_upper_env}
\title{Top of Second Trace Envelope}
\usage{
.bottom_upper_env(rolledImage, max_roc = 25, sepDist = 10, maxNoise = 100)
}
\arguments{
\item{rolledImage}{Image that has been put through the .roll_image()}

\item{max_roc}{maximum rate of change allowed between two pixels on the line before deemed as noise}

\item{sepDist}{how far you want the envelope to be above the line you are tracing}

\item{maxNoise}{the length of creating points alowed before considered to be off of the trace}
}
\value{
vector of points for a line, in the correct scaling for the rolled image.
can do nrow(imageMatrix) - bottomcut + Envelope change the scaling
}
\description{
Top of Second Trace Envelope
}
