% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.dir_Str}
\alias{.dir_Str}
\title{Directory Structure Creation of image year}
\usage{
.dir_Str(imageName, pathToWorkingDir)
}
\arguments{
\item{imageName}{A string of the image name}

\item{pathToWorkingDir}{Where the digitized images will be put}
}
\value{
The pwd for that specific image
}
\description{
Creates the directory structure for the digitized images along with setting with
pwd for that specific image, if that year doesn't exist, it will create one
}
