% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.dir_Str_Improvement}
\alias{.dir_Str_Improvement}
\title{Directory Structure Creation of image name}
\usage{
.dir_Str_Improvement(imageName, pathToWorkingDir)
}
\arguments{
\item{imageName}{A string of the image name}

\item{pathToWorkingDir}{Where the digitized images will be put}
}
\value{
The pwd for that specific image
}
\description{
Creates the directory structure for the digitized images along with setting with
pwd for that specific image, if a dir doesn't exist for the image name, it will create one
}
