% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.edge_peaks_rm}
\alias{.edge_peaks_rm}
\title{Scanning Flair Around Outside of Image}
\usage{
.edge_peaks_rm(FindingPeaksdf, rowSums, percentEdge, percentEdgeForLeft = NULL)
}
\arguments{
\item{FindingPeaksdf}{the resulting data frame from finding_peaks
(if called separately make sure in correct form nx2: Index column and Height column)}

\item{rowSums}{of the imported matrix from original picture}

\item{percentEdge}{passed into .allowed_edge_distance}

\item{percentEdgeForLeft}{is user needs to specify a different percentage for left and right side cutoff
will make percentEdge for the right side and percentEdgeForLeft for the left side of image}
}
\value{
void
}
\description{
Checks the list of peaks for any flairs in the photos at start and end, they are removed
}
