% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.envelope_creation}
\alias{.envelope_creation}
\title{Envelope Creation For a Trace}
\usage{
.envelope_creation(
  rolledImage,
  chosenRoughBound,
  max_roc = 25,
  sepDist = 20,
  topOfLine = TRUE,
  maxNoise = 100
)
}
\arguments{
\item{rolledImage}{Image that has been put through the .roll_image()}

\item{chosenRoughBound}{vector of the rough bound (use apply on the columns)}

\item{max_roc}{maximum rate of change allowed between two pixels on the line before deemed as noise}

\item{sepDist}{how far you want the envelope to be below the line you are tracing}

\item{topOfLine}{bool, if you are tracing the top of the bottom of the line}

\item{maxNoise}{the length of creating points alowed before considered to be off of the trace}
}
\value{
corected bound for the envelope
}
\description{
Takes a rough est for the envelope and removes any unwanted noise
}
