% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.envelopegapfiller}
\alias{.envelopegapfiller}
\title{envelope gap filling for TISI}
\usage{
.envelopegapfiller(x, y, nCol = NULL)
}
\arguments{
\item{x}{vector of x values}

\item{y}{vector of y values}

\item{nCol}{the length out that you want}
}
\value{
data.frame of x and y values
}
\description{
envelope gap filling for TISI
}
