% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.find_a_number}
\alias{.find_a_number}
\title{Find A Specific Number}
\usage{
.find_a_number(vector, specNumber)
}
\arguments{
\item{vector}{a generic vector of any length}

\item{specNumber}{a number that you would like to find all sequences of}
}
\value{
vector of all occurrences lengths of each sequence and the index of each start
}
\description{
Used for finding all sequences of one number in a vector and start indexes
and lengths of each sequence are returned to the user
}
