% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.finding_distance_to_peaks}
\alias{.finding_distance_to_peaks}
\title{Distance to Closest Peak or Border}
\usage{
.finding_distance_to_peaks(FindingPeaksdf, rowSums, peakIndex)
}
\arguments{
\item{FindingPeaksdf}{the resulting data frame from finding_peaks
(if called separately make sure in correct form nx2: Index column and Height column)}

\item{rowSums}{of the imported matrix from original picture}

\item{peakIndex}{the index of FindingPeaksdf that contains the peak you want to look at}
}
\value{
dataframe  1x2 leftDist, rightDist for columns
}
\description{
Finds the closest peak(or border) index for a given peak on either side
}
