% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.finding_Peak_Start_Ends}
\alias{.finding_Peak_Start_Ends}
\title{Finding starting point and ending point for each peak}
\usage{
.finding_Peak_Start_Ends(FindingPeaksdf, rowSums)
}
\arguments{
\item{FindingPeaksdf}{the resulting data frame from finding_peaks
(if called separately make sure in correct form nx2: Index column and Height column)}

\item{rowSums}{of the imported matrix from original picture}
}
\value{
FindingPeaksdf along with the starting and ending points of each peak
}
\description{
Finding starting point and ending point for each peak
}
