% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.for_bright_image}
\alias{.for_bright_image}
\title{Bright image scaling to gaussian}
\usage{
.for_bright_image(
  imageMatrix,
  Filter = c(13, 13),
  method = "delation",
  threshold = 0.8
)
}
\arguments{
\item{imageMatrix}{An imported image, can be imported with tiff_import()}

\item{Filter}{a vector specifying the dimensions of the kernel,
which will be used to perform either delation or erosion, such as c(13,13)}

\item{method}{one of 'delation'(adds to image, making brights brighter), 'erosion' (subtracts from image brights darker)}

\item{threshold}{should be between 0 and 1 for normalized images}
}
\value{
Gaussian matrix scaled for bright
}
\description{
Bright image scaling to gaussian
}
