% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.get_image_year}
\alias{.get_image_year}
\title{Get Year From Image Name}
\usage{
.get_image_year(imageName)
}
\arguments{
\item{imageName}{a string}
}
\value{
Year
}
\description{
Get Year From Image Name
}
