% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.get_trace_start_ends}
\alias{.get_trace_start_ends}
\title{Trace Starting and Ending Points}
\usage{
.get_trace_start_ends(
  imageMatrix,
  cutPercentage = 1,
  peakThreshold = 5,
  gapAllow = 20,
  returnMat = TRUE,
  maxStart = 700,
  minEnd = 4800
)
}
\arguments{
\item{imageMatrix}{Imported image into matrix form, can use tiff_import(), processed already}

\item{cutPercentage}{A bound for which the start and end is never found, could be a flare
on the sides of the scanned image}

\item{peakThreshold}{Smallest difference allowed for a difference to be a possible peak}

\item{gapAllow}{Distance in which another index could be considered to be the same peak,
applicable for large peaks}

\item{returnMat}{bool, default is to return the image with removed start and end to the user, if FALSE
will return just the index of the start and end}

\item{maxStart}{absolute max for the left side of the image (can be overridden with the length)}

\item{minEnd}{absolute min for the right side of the image ( can be overridden with the 0)}
}
\value{
Image matrix with removed start and end, or the index of these start and ends
}
\description{
Can also be used to remove those ranges where the trace isn't if returnMat is TRUE
}
