% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{.hdv_check}
\alias{.hdv_check}
\title{HDV Checking}
\usage{
.hdv_check(imageName)
}
\arguments{
\item{imageName}{The file name for the specific image}
}
\value{
warning if these are found (use a tryCatch)
}
\description{
HDV is a type of image with three normal images on the same scan, these cause big problems
filters them out by the name
}
