% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.highest_peaks}
\alias{.highest_peaks}
\title{Highest Peaks}
\usage{
.highest_peaks(FindingPeaksdf, maxPeaksAllowed)
}
\arguments{
\item{FindingPeaksdf}{the resulting data frame from finding_peaks
(if called separately make sure in correct form nx2: Index column and Height column)}

\item{maxPeaksAllowed}{The maximum number of lines on a picture you would like to find}
}
\value{
The highest peaks up to the max number allowed
}
\description{
Finds the highest peaks up to the maximum number of peaks allowed(set by user on call)
}
