% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{.horizontal_image_check}
\alias{.horizontal_image_check}
\title{Horizontal Image Checking}
\usage{
.horizontal_image_check(matrix)
}
\arguments{
\item{matrix}{The matrix of an imported tiff,png,etc...}
}
\value{
The matrix horizontally
}
\description{
Checks to see if the image is vertical or horizontal, if vertical, makes horizontal
}
