% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{.is_tiff}
\alias{.is_tiff}
\title{Tiff File Check}
\usage{
.is_tiff(character)
}
\arguments{
\item{character}{in the form something.tif}
}
\value{
bool of TRUE or FALSE
}
\description{
checks to see if the last part of a file name is .tif or .tiff
}
\examples{
.is_tiff(testing.tif)
}
