% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.not_bright_image}
\alias{.not_bright_image}
\title{Non bright image scaling with gaussian deconvolution}
\usage{
.not_bright_image(
  imageMatrix,
  Filter = c(8, 8),
  method = "delation",
  threshold = 0.5
)
}
\arguments{
\item{imageMatrix}{An imported image, can be imported with tiff_import()}

\item{Filter}{a vector specifying the dimensions of the kernel,
which will be used to perform either delation or erosion, such as c(8,8)}

\item{method}{one of 'delation'(adds to image, making brights brighter), 'erosion' (subtracts from image brights darker)}

\item{threshold}{should be between 0 and 1 for normalized images}
}
\value{
Gaussian matrix scaled for bright
}
\description{
Non bright image scaling with gaussian deconvolution
}
