% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{.null}
\alias{.null}
\title{Null Argument Checking}
\usage{
.null(parameter, parameterName)
}
\arguments{
\item{parameter}{what you would like to check if is null}

\item{parameterName}{Allows the error message to be in context}
}
\value{
stop if is null
}
\description{
Checks a parameter for a null argument
}
