% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.plot_peaks}
\alias{.plot_peaks}
\title{Plotting Peaks from find_peaks}
\usage{
.plot_peaks(FindPeaksdf, rowSums, StartEndLine = TRUE)
}
\arguments{
\item{FindPeaksdf}{usually found with find_peaks nx4: Index, Height, PeakStart, PeakEnd}

\item{rowSums}{of the imported matrix from original picture}

\item{StartEndLine}{bool, if you want a plotted starting and ending line for each peak}
}
\value{
void
}
\description{
This can be used alone from find_peaks, ensure that you have the correct df form.
starting are green lines, ending are blue lines
}
