% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiddenFunctions.R
\name{.process_image}
\alias{.process_image}
\title{Process Image}
\usage{
.process_image(
  imageMatrix,
  FilterBright = c(13, 13),
  FilterNonBright = c(8, 8),
  methodBright = "delation",
  methodNonBright = "delation",
  thresholdBright = 0.8,
  thresholdNonBright = 0.5,
  beta0 = -2.774327,
  beta1 = 51.91687,
  cutoffProbability = 0.5,
  NADefault = 0
)
}
\arguments{
\item{imageMatrix}{Imported image into matrix form, can use tiff_import()}

\item{FilterBright}{Vector specifying the dimensions of the kernel,
which will be used to perform either delation or erosion, such as c(13,13)}

\item{FilterNonBright}{Vector specifying the dimensions of the kernel,
which will be used to perform either delation or erosion, such as c(8,8)}

\item{methodBright}{one of 'delation'(adds to image, making brights brighter), 'erosion' (subtracts from image brights darker)}

\item{methodNonBright}{one of 'delation'(adds to image, making brights brighter), 'erosion' (subtracts from image brights darker)}

\item{thresholdBright}{should be between 0 and 1 for normalized images Default = 0.8}

\item{thresholdNonBright}{should be between 0 and 1 for normalized images Default = 0.5}

\item{beta0}{From logistic regression on what images to be considered bright}

\item{beta1}{From logistic regression on what images to be considered bright}

\item{cutoffProbability}{Passed into bright: The probability cut off for the decision of an imageMatrix being bright}

\item{NADefault}{The defult value set to points of NA found by the system}
}
\value{
The processed image with the gaussian and the non gaussian in an array labeled respectively
}
\description{
Process Image
}
