% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{env_start_end}
\alias{env_start_end}
\title{Envelope Starts and Ends}
\usage{
env_start_end(
  traceMatrix,
  thresh = 300,
  gapLengthCutoff = 20,
  returnMatrix = TRUE
)
}
\arguments{
\item{traceMatrix}{a single isolated trace, with only one trace(no timing on the plot)}

\item{thresh}{How far it will go before deciding were already past the numbers at start of traces}

\item{gapLengthCutoff}{How Large the gap between the text and the trace must be in order to consider that
point the new start, if gap is less then cutoff, will keep looking until it gets to the thresh value}

\item{returnMatrix}{Will return the cut matrix if TRUE, if FALSE will return list of the Start and End}
}
\value{
list of the Start and End or cut matrix
}
\description{
A better version of .get_image_starts_ends but meant for single traces that
are isolated on their own picture.
}
