% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{find_cuts}
\alias{find_cuts}
\title{Find Cut Positions}
\usage{
find_cuts(
  imageMatrix,
  cutPercentage = 2,
  percentFromEdge = 2,
  percentEdgeForLeft = NULL,
  shortestAllowedSeqOfZeros = 25
)
}
\arguments{
\item{imageMatrix}{imported with import_process_image() or an equivilent}

\item{cutPercentage}{A bound for which the start and end is never found, could be a flare
on the sides of the scanned image}

\item{percentFromEdge}{used in find_peaks if you know there wont be a peak
in a region}

\item{percentEdgeForLeft}{passed into find peaks, if not specified (NULL), uses
percentFromEdge for both left and right sides, if specified, percentFromEdge
is defaulted to just the right side of the plot}

\item{shortestAllowedSeqOfZeros}{smallest gap allowed to be found to consider
the trace not intersecting the timing marks in pixels}
}
\value{
list of the top and bottom cuts, warnings are returned if exist
}
\description{
Attempts to find a straight line between the timing marks and the traces(.bottom_image_cut)
and a straight line between the traces and the writing at the top of the images(.top_image_cut)
}
