% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{find_envelopes}
\alias{find_envelopes}
\title{Finding Envelopes for Two Traces}
\usage{
find_envelopes(
  imageMatrix,
  rolledImage,
  bottomCut,
  topCut,
  returnType,
  sepDist = 10,
  max_roc = 25,
  maxNoise = 100,
  improveTTopEnvelope = data.frame(x = NA, y = NA),
  improveBTopEnvelope = data.frame(x = NA, y = NA),
  improveTBottomEnvelope = data.frame(x = NA, y = NA),
  improveBBottomEnvelope = data.frame(x = NA, y = NA),
  trimTop = 0,
  trimBottom = 0
)
}
\arguments{
\item{imageMatrix}{ImageMatrix from import_process_image()}

\item{rolledImage}{the image rolled by mean_roll_image()}

\item{bottomCut}{a point that is the height of a horizontal line between the timing marks
and the bottom trace}

\item{topCut}{Top cutoff between the date writing and the top trace}

\item{returnType}{Either "MatrixScaled" used if you are manipulating with imageMatrix,
"PlottingScaled" used if you are plotting the imageMatrix with overlay of these lines
"RolledImageScaled" used if you are overlaying the rolled image with these lines}

\item{sepDist}{how far you want the envelope to be below the line you are tracing}

\item{max_roc}{maximum rate of change allowed between two pixels on the line before deemed as noise}

\item{maxNoise}{the length of creating points allowed before considered to be off of the trace}

\item{improveTTopEnvelope}{data.frame of col x and col y representing points for top of the top envelope pair
doesn't need to be a point for each column in original picture, will fill in the blanks}

\item{improveBTopEnvelope}{data.frame of col x and col y representing points for Bottom of the top envelope pair
doesn't need to be a point for each column in original picture, will fill in the blanks}

\item{improveTBottomEnvelope}{data.frame of col x and col y representing points for top of the bottom envelope pair
doesn't need to be a point for each column in original picture, will fill in the blanks}

\item{improveBBottomEnvelope}{data.frame of col x and col y representing points for bottom of the bottom envelope pair
doesn't need to be a point for each column in original picture, will fill in the blanks}

\item{trimTop}{how much has been trimmed by TISI or TIS}

\item{trimBottom}{how much has been trimmed by TISI or TIS}
}
\value{
list of all four envelopes
}
\description{
Finding Envelopes for Two Traces
}
