% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{find_paths_for_keyword}
\alias{find_paths_for_keyword}
\title{Searching For Keyword}
\usage{
find_paths_for_keyword(path = "~/", keyword)
}
\arguments{
\item{path}{Where you would like the function to look.}

\item{keyword}{The word that you are trying to find in file names.}
}
\value{
Vector with all paths for each item found with the keyword contained in it.
}
\description{
Finds all files that contain a specific string, in this case for finding files
that contain a specific year. Default is to also look in symlinks for files, if a dir is not accessable
will return an warning to the user
}
