% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{find_peaks}
\alias{find_peaks}
\title{Finding Peaks for Image}
\usage{
find_peaks(
  rowSums,
  minDistance,
  maxPeakNumber,
  percentFromEdge,
  percentEdgeForLeft = NULL,
  minPeakHeight = (0.05 * max(rowSums)),
  plots = TRUE,
  StartEndplotLine = TRUE
)
}
\arguments{
\item{rowSums}{of the imported matrix from original picture}

\item{minDistance}{Minimum distance aloud between found peaks}

\item{maxPeakNumber}{Maximum peaks you are trying to find}

\item{percentFromEdge}{Percentage of the picture that you would like removed from edge
due to an over exposed picture or flares in scanning
if percentageLeftSide is unspecified uses percentage for both i.e left = 1 - right}

\item{percentEdgeForLeft}{the left percentage that you want to remove(default = NULL)}

\item{minPeakHeight}{minimum peak height required for a peak to be counted
Default is 5perc of the max(rowSums)}

\item{plots}{Plots the peaks on rowSums}

\item{StartEndplotLine}{If you want to see lines at start and ends of each peaks
(starts are green, ends are blue)}
}
\value{
data frame of peaks, peak heights, starts and ends for each peak
}
\description{
Finds peaks in the row sums of an image matrix, used for finding starts of lines
that are horizontal
}
