% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{import_process_image}
\alias{import_process_image}
\title{Import and Process a .tiff or .tif image}
\usage{
import_process_image(
  imageName,
  file_loc,
  trimAmountTop = 100,
  trimAmountBottom = 50,
  beta0 = -2.774327,
  beta1 = 51.91687,
  cutoffProbability = 0.5,
  NADefault = 0,
  FilterBright = c(13, 13),
  FilterNonBright = c(8, 8),
  methodBright = "delation",
  methodNonBright = "delation",
  thresholdBright = 0.8,
  thresholdNonBright = 0.5
)
}
\arguments{
\item{imageName}{The name of the file}

\item{file_loc}{the path from ~/ to the dir where the file is}

\item{trimAmountTop}{Number of pixels off of top of image (usually for common flares)}

\item{trimAmountBottom}{Number of pixels off of bottom of image}

\item{beta0}{The intercept of a logistic regression default is for magnetograms}

\item{beta1}{The slope of the logistic regression default is for magnetograms}

\item{cutoffProbability}{The probability cut off for the decision of an imageMatrix being bright}

\item{NADefault}{The default value set to NA's found in the matrix}

\item{FilterBright}{Vector specifying the dimensions of the kernel,
which will be used to perform either delation or erosion, such as c(13,13)}

\item{FilterNonBright}{Vector specifying the dimensions of the kernel,
which will be used to perform either delation or erosion, such as c(8,8)}

\item{methodBright}{one of 'delation'(adds to image, making brights brighter), 'erosion' (subtracts from image brights darker)}

\item{methodNonBright}{one of 'delation'(adds to image, making brights brighter), 'erosion' (subtracts from image brights darker)}

\item{thresholdBright}{should be between 0 and 1 for normalized images Default = 0.8}

\item{thresholdNonBright}{should be between 0 and 1 for normalized images Default = 0.5}
}
\value{
imageMatrix processed and trimed, in landscape orientation
}
\description{
Import and Process a .tiff or .tif image
}
