% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{isolate_traces}
\alias{isolate_traces}
\title{Isolation of Two Traces}
\usage{
isolate_traces(
  imageMatrix,
  topEnvelope,
  topLowerEnvelope,
  bottomUpperEnvelope,
  bottomEnvelope
)
}
\arguments{
\item{imageMatrix}{The processed Image matrix with import_process_image()
and trimmed if applicable}

\item{topEnvelope}{Top envelope for the top trace (scaled to your matrix correctly)}

\item{topLowerEnvelope}{Bottom envelope for the top trace (scaled to your matrix correctly)}

\item{bottomUpperEnvelope}{The top envelope for the second trace (scaled to your matrix correctly)}

\item{bottomEnvelope}{Bottom envelope for the bottom trace (scaled to your matrix correctly)}
}
\value{
list of both isolated trace matrices
}
\description{
Uses envelopes to create two different matrices with one trace on each
}
\details{
NOTE: make sure that your envelopes are scaled to your matrix
}
