% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{mean_roll_image}
\alias{mean_roll_image}
\title{Image Blurring with Rollmean}
\usage{
mean_roll_image(imageMatrix, topcut, bottomcut, fill = "extend", k = 40)
}
\arguments{
\item{imageMatrix}{the image matrix from import_process_image}

\item{topcut}{Top cutoff from .top_cut()}

\item{bottomcut}{Bottom cutoff from .bottom_cut()}

\item{fill}{Default "extend" see rollMean in zoo for details}

\item{k}{See rollMean() in zoo for details}
}
\value{
rolled image matrix to user
}
\description{
uses rollMean from zoo to blurr image ( used to get rid of timing gaps)
}
