% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{not_empty_file}
\alias{not_empty_file}
\title{Empty File Checking}
\usage{
not_empty_file(filePath, fileName)
}
\arguments{
\item{filePath}{The path for the directory where the file is located}

\item{fileName}{The name of the file located in the path directory}
}
\value{
a bool of TRUE or FALSE
}
\description{
Checks a given file path with name to see if the file is empty 0b
}
