% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{plot_success}
\alias{plot_success}
\title{Plotting Successful Digitization}
\usage{
plot_success(
  imageMatrix,
  rolledImage,
  topCut,
  bottomCut,
  topStartEnds,
  bottomStartEnds,
  topTrace,
  bottomTrace,
  plotEnv,
  maxNoise = 250,
  max_roc = 35,
  sepDist = 10,
  pathToWorkingDir = "~/",
  imageName,
  intersectionFlag = FALSE
)
}
\arguments{
\item{imageMatrix}{Matrix for a certain image (scaled)}

\item{rolledImage}{Using mean_roll_image}

\item{topCut}{A point which indicates a horizontal line between the top text
and the first trace}

\item{bottomCut}{A point which indicates a horizontal line between the bottom trace
and the timing marks (if a line can be found)}

\item{topStartEnds}{Top trace start and ends for isolated trace}

\item{bottomStartEnds}{Bototm trace start and ends for isolated trace}

\item{topTrace}{Top trace matrix for the isolated trace}

\item{bottomTrace}{Bottom trace matrix for the isolated trace}

\item{plotEnv}{The envelope that you would like to put onto the plot}

\item{maxNoise}{The length of creating points allowed before considered to be off of the trace}

\item{max_roc}{Maximum rate of change allowed between two pixels on the line before deemed as noisee}

\item{sepDist}{How far you want the envelope to be below the line you are tracing}

\item{pathToWorkingDir}{Which directory you want the image to end up in}

\item{imageName}{The name of the image (usually the file name)}

\item{intersectionFlag}{If intersection is known to user, or is found with TISI or TIS,
will not plot full, will consider it to be a failed process plot}
}
\value{
void
}
\description{
Plotting Successful Digitization
}
