% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{plot_with_warnings}
\alias{plot_with_warnings}
\title{Plot for Fail Processed Images}
\usage{
plot_with_warnings(imageMatrix, topCut, bottomCut, pathToWorkingDir, imageName)
}
\arguments{
\item{imageMatrix}{Matrix for an image(scaled)}

\item{topCut}{A point which indicates a horizontal line between the top text
and the first trace}

\item{bottomCut}{A point which indicates a horizontal line between the bottom trace
and the timing marks (if a line can be found)}

\item{pathToWorkingDir}{Which directory you want the image to end up in}

\item{imageName}{The file name of the image}
}
\value{
void
}
\description{
Plot for Fail Processed Images
}
