% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{spike_check}
\alias{spike_check}
\title{Trace Spike Check}
\usage{
spike_check(trace, spikeThreshold = 50)
}
\arguments{
\item{trace}{a single trace}

\item{spikeThreshold}{how big the difference has to be to be considered to be a spike}
}
\value{

}
\description{
Looks for abnormal spikes in the trace that wasn't smoothed out before
}
