% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{tiff_import}
\alias{tiff_import}
\title{Importing Tiff Files}
\usage{
tiff_import(fileName, fileLoc)
}
\arguments{
\item{fileName}{The image name}

\item{fileLoc}{the path from ~/ to the dir where the file is}
}
\description{
Reads tiff also checks for if the image is a .tif or .tiff using .is_tiff()
}
