% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{trim_sides}
\alias{trim_sides}
\title{Rough Trim for Sides}
\usage{
trim_sides(image, trimAmountLeft = 2, trimAmountRight = 2)
}
\arguments{
\item{image}{A matrix}

\item{trimAmountLeft}{in percentage}

\item{trimAmountRight}{in percentage}
}
\value{
the matrix without the specified bounds
}
\description{
trims the right and left sides if the user knows where there will definitely not be any points
}
