% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{trim_top_bottom}
\alias{trim_top_bottom}
\title{Trim on top and bottom of the image in pixels for consistency}
\usage{
trim_top_bottom(image, trimAmountTop = 100, trimAmountBottom = 50)
}
\arguments{
\item{image}{A matrix}

\item{trimAmountTop}{Number of pixels}

\item{trimAmountBottom}{Number of pixels}
}
\value{
the matrix without the specified bounds
}
\description{
trims  the top and bottom if the user knows where there will definitely not be any points
}
