% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userFunctions.R
\name{triple_check}
\alias{triple_check}
\title{Triple Trace Checking}
\usage{
triple_check(
  imageMatrix,
  topCut,
  bottomCut,
  minDistance = 50,
  percentFromEdge = 2,
  thresholdHeight = 200,
  thresholdDistance = 250,
  threshCutImage = 500
)
}
\arguments{
\item{imageMatrix}{Horizontal image processed}

\item{topCut}{from .topCut() line between the text and the two traces}

\item{bottomCut}{from .bottomCut() line after traces before the two timing marks}

\item{minDistance}{The min distance aloud between found peaks}

\item{percentFromEdge}{the distance alows from the edge of the image}

\item{thresholdHeight}{How separated the heights of the three peaks can be}

\item{thresholdDistance}{How far the timing lines can be from each other(
saves the cases when the peaks are same height as the timing traces)}

\item{threshCutImage}{how much different the peak heights have to be in the cut image
for the image to be considered to be a triple.}
}
\value{
TRUE or FALSE for finding a triple or not
}
\description{
Using find_peaks, checks for three traces on the same Image, These can't be done
currently by this alg
}
